APIBakeryMgmt/requirement.txt
Django==3.1.4
djangorestframework==3.12.2
postgres==3.0.0W
gunicorn==20.0.4
whitenoise==5.2.0